<한다영> 관리자 페이지 세션처리, 질문 등록 버그 수정 (중간푸시)
- 관리자 세션처리: 페이지 접근시 로그인 하지 않은 사용자나 일반 유저는 접근 차단
- 관리자 사이드바 통계랑 문의 내역 사이에 구분자 추가
- 질문 등록시 질문 내용 받지 못했던 오류 수정

+++ codeblue/src/main/java/com/example/codeblue/controller/AdminController.java
@@ -1,5 +1,7 @@
 package com.example.codeblue.controller;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,77 +15,97 @@ public class AdminController {
 	
 	//삭제된 게시물 리스트 페이지
 	@GetMapping("/adminWithdrawQuestionBoard")
-	public String getWithdramQuestionBoard() {
+	public String getWithdramQuestionBoard(HttpSession session) {
 		System.out.println("::: get - getWithdramQuestionBoard :::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/adminWithdrawQuestionBoard";
 	}
-	// 코드블루 소개 패이지
-	@GetMapping("index")
-	public String getIndex() {
-		return "/index";
-	}
 	
-	// 사이트 하단 부 
-	@GetMapping("footer")
-	public String getFooter() {
-		return "/codeblue/resource/footerCopy";
-	}
-	
-	// 사이트 관리자 메뉴
-	@GetMapping("sideBar")
-	public String getSideBar() {
-		return "/codeblue/admin/resource/sidebarCopy";
-	}
 	//FAQ 상세 페이지
 	@GetMapping("/adminFaqOne")
-	public String getFaqOne() {
+	public String getFaqOne(HttpSession session) {
 		System.out.println(":::get - getFaqOne:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return"/codeBlue/admin/adminFaqOne";
 	}
 	//FAQ 수정페이지
 	@GetMapping("/adminFaqModify")
-	public String modifyFaq() {
+	public String modifyFaq(HttpSession session) {
 		System.out.println(":::get - modifyFaq:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return"/codeBlue/admin/adminModifyFaqBoard";
 	}
 	//FaQ 게시판 페이지로 이동
 	@GetMapping("/adminFaqBoard")
-	public String getFaqBoard() {
+	public String getFaqBoard(HttpSession session) {
 		System.out.println(":::get - getFaqBoard:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/adminFaqBoard";
 	}
 	//FaQ 게시판 페이지로 이동
 	@GetMapping("/adminFaqBoardAdd")
-	public String addFaqBoard() {
+	public String addFaqBoard(HttpSession session) {
 		System.out.println(":::get - addFaqBoard:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/adminAddFaqBoard";
 	}
 	//병원정보 수정
 	@GetMapping("/modifyHospital")
-	public String getModifyHospital(@RequestParam(value="hospitalId")int hospitalId) {
+	public String getModifyHospital(HttpSession session, @RequestParam(value="hospitalId")int hospitalId) {
 		System.out.println(":::get - getModifyHospital:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		System.out.println("hospital"+hospitalId);
 		return "/codeBlue/admin/adminModifyHospital";
 	}
 	
 	//제휴병원 게시판
 	@GetMapping("/adminHospital")
-	public String getHospital() {
+	public String getHospital(HttpSession session) {
 		System.out.println(":::get - getHospital:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return"/codeBlue/admin/adminHospital";
 	}
 	
 	//제휴병원 입력
 	@GetMapping("/addHospital")
-	public String getAddHospital() {
+	public String getAddHospital(HttpSession session) {
 		System.out.println(":::get - AddHospital:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return"/codeBlue/admin/adminAddHospital";
 	}
 	
 	//제휴병원 삭제
 	@GetMapping("/removeHospital")
-	public String postRemoveHospital(@RequestParam(value="hospitalId")int hospitalId) {
+	public String postRemoveHospital(HttpSession session, @RequestParam(value="hospitalId")int hospitalId) {
 		System.out.println(":::post - postRemoveHospital:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		System.out.println("hospitalId"+hospitalId);
 		adminService.removeHospital(hospitalId);
 		return "/codeBlue/admin/adminHospitalOne";
@@ -91,69 +113,113 @@ public class AdminController {
 	
 	//제휴병원 상세보기
 	@GetMapping("/getHospitalOne")
-	public String getHospitalOne(@RequestParam(value="hospitalId")int hospitalId) {
+	public String getHospitalOne(HttpSession session, @RequestParam(value="hospitalId")int hospitalId) {
 		System.out.println(":::get - getHospitalOne:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		System.out.println("hospitalId"+hospitalId);
-		
 		return"/codeBlue/admin/adminHospitalOne";
 	}
+	
 	//신고내역 리스트 페이지로 이동
 	@GetMapping("adminReportBoard")
-	public String getReport() {
+	public String getReport(HttpSession session) {
 		System.out.println("::: get - getReport :::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/report";
 	}
 	//문의내역 리스트 페이지로 이동
 	@GetMapping("adminInquiryBoard")
-	public String getInquiry() {
+	public String getInquiry(HttpSession session) {
 		System.out.println("::: get - adminInquiry :::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/inquiry";
 	}
 	//문의내역 상세 페이지로 이동
 	@GetMapping("adminInquiryOne")
-	public String getInquiryOne() {
+	public String getInquiryOne(HttpSession session) {
 		System.out.println("::: get - getInquiryOne :::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/inquiryOne";
 	}
 	//신고내역 상세 페이지로 이동
 	@GetMapping("adminReportOne")
-	public String getReportOne() {
+	public String getReportOne(HttpSession session) {
 		System.out.println("::: get - getReportOne :::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/reportOne";
 	}
 	
 	@GetMapping("/adminUserList")
-	public String adminUserList() {
+	public String adminUserList(HttpSession session) {
 		System.out.println("::: get - adminUserList :::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/adminUserList";
 	}
 	
 	@GetMapping("/adminNoticeBoard")
-	public String getNotice() {
+	public String getNotice(HttpSession session) {
 		System.out.println(":::get - getNotice:::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 	return "/codeBlue/admin/adminNoticeBoard";
 	}
 	
 	@GetMapping("/adminNoticeOne")
-	public String getNoticeOne() {
+	public String getNoticeOne(HttpSession session) {
 		System.out.println(":::get - getNoticeOne");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/adminNoticeOne";
 	}
 	
 	@GetMapping("/adminNoticeAdd")
-	public String getNoticeAdd() {
+	public String getNoticeAdd(HttpSession session) {
 		System.out.println(":::get - getNoticeAdd");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return"/codeBlue/admin/adminAddNotice";
 	}
 	//질문 리스트 가져오기
 	@GetMapping("/adminQuestionBoard")
-	public String getQuestionBoardList() {
+	public String getQuestionBoardList(HttpSession session) {
 		System.out.println("::: get - getQuestionBoardList :::");
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/adminQuestionBoard";
 	}
 	
 	@GetMapping("admin")
-	public String getAdmin() {
+	public String getAdmin(HttpSession session) {
+		if(session.getAttribute("authority")==null || session.getAttribute("authority").equals("user")) {
+			System.out.println("user admin 페이지 접근 시도 차단");
+			return "/home";
+		}
 		return "/codeBlue/admin/home";
 	}
 }
